# backend-go/api/openapi.yaml
openapi: 3.0.3
info:
  title: Creation Agent Studio API
  version: 2.0.0
  description: |
    Creation Agent Studio backend contract (Go control plane).

    This file is the single source of truth for the HTTP API.
    Go server types and TypeScript clients are generated from it.

    ## Error envelope conventions (legacy-compatible)

    The React frontend was validated against the Django reference
    implementation, which produced four distinct error shapes:

    - `{"detail": "message"}` — 403/404 style errors (DRF PermissionDenied / NotFound).
    - `{"error": "message"}` — simple JSON errors (identity endpoints, 409/502 conflicts).
    - `{"field": ["message", ...]}` — field validation errors (DRF serializer style).
    - `["message"]` — bare array, 400 for hand-raised DRF ValidationErrors (e.g. "application_id is required").

    All four shapes are reproduced exactly. Clients should read the
    `detail` / `error` / first field message in that order.

    ## Authentication

    Normal users authenticate with the HttpOnly opaque session cookie
    `studio_session` (set by OAuth exchange / admin login). All
    state-changing requests additionally require the CSRF header
    `X-CSRF-Token` matching the `studio_csrf` cookie (double-submit).
    `GET /api/v2/artifacts/{id}/open` is cookie-authenticated because it is
    opened via browser navigation (window.open) and cannot send headers.

    ## SSE

    `GET /api/v2/runs/{id}/stream` responds with `text/event-stream`.
...
servers:
...
paths:
...
  /api/identity/oauth/start:
    get:
      operationId: oauthStart
      tags: [identity]
      summary: Redirect (302) the browser to the Feishu OAuth authorize page
      parameters:
        - name: return_to
          in: query
          schema: { type: string, default: "/" }
          description: SPA path to return to after login; must start with "/" and not "//".
      responses:
        "302":
          description: Redirect to Feishu authorize URL (state signed, 10 min TTL).
          headers:
            Location:
              schema: { type: string }
        "500":
          description: OAuth not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }
  /api/identity/oauth/exchange:
    get:
      operationId: oauthExchange
      tags: [identity]
      summary: Exchange the OAuth ?code&state for a studio session
      description: |
        Verifies the signed state, exchanges the code at the Feishu token
        endpoint, upserts the local user (auth_source=feishu), stores the
        refresh token AES-256-GCM encrypted, sets the `studio_session`
        HttpOnly cookie and returns the session user plus JWT tokens for the
        legacy transition window.
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Authenticated
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [user, tokens, return_to]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  tokens: { $ref: "#/components/schemas/Tokens" }
                  return_to: { type: string }
        "400":
          description: missing code / invalid or expired oauth state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }
        "502":
          description: oauth exchange failed / no access token / user info failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }
...
  /api/identity/admin/login:
    post:
      operationId: adminLogin
      tags: [identity]
      summary: Local admin login (Argon2id password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Admin session established
          content:
            application/json:
              schema:
                type: object
                required: [id, username, is_staff, auth_source]
                properties:
                  id: { type: integer }
                  username: { type: string }
                  is_staff: { type: boolean }
                  auth_source: { type: string }
                  display_name: { type: string }
                  display_id: { type: string }
        "401":
          description: invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }

  # ─────────────────────────────────────────────────────────────── auth ──
  /api/auth/login/:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Local username/password login (transition window)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Logged in; also sets studio_session cookie
          content:
            application/json:
              schema:
                type: object
                required: [user, tokens]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  tokens: { $ref: "#/components/schemas/Tokens" }
        "401":
          description: bad credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
...
  /api/auth/token/refresh/:
    post:
      operationId: tokenRefresh
      tags: [auth]
      summary: Rotate a JWT refresh token (transition window)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh]
              properties:
                refresh: { type: string }
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access: { type: string }
                  refresh: { type: string }
...
  /api/v2/applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List openable applications (bare array, no pagination envelope)
      parameters:
        - name: kind
          in: query
          schema: { type: string, enum: [chat, task, custom, all], default: chat }
          description: Invalid values fall back to chat.
        - name: scope
          in: query
          schema: { type: string, enum: [public, mine, manage], default: public }
          description: Invalid values fall back to public.
        - name: include_unbound
          in: query
          schema: { type: string }
          description: Truthy when case-insensitively in {1,true,yes}.
      responses:
        "200":
          description: Application list items (bare array)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApplicationListItem" }
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create an application (admin only — agent marketplace "new Aily agent")
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationCreateRequest" }
      responses:
        "201":
          description: Created (compact detail shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "400":
          description: field errors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FieldErrors" }
        "403":
          description: only admins may add agents
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/{id}:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    get:
      operationId: getApplication
      tags: [applications]
      summary: Application detail (compact shape)
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "404":
          description: application not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update an application / rebind runtime (creator or staff)
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationUpdateRequest" }
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "403":
          description: only creator or staff
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "404":
          description: application not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application (409 when referenced by conversations/runs)
      responses:
        "204":
          description: Deleted
        "403":
          description: only creator or staff
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "409":
          description: has conversations or runs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/{id}/favorite:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      operationId: favoriteApplication
      tags: [applications]
      summary: Pin application (idempotent)
      responses:
        "200":
          description: Favorite state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FavoriteResult" }
        "404":
          description: application not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    delete:
      operationId: unfavoriteApplication
      tags: [applications]
      summary: Unpin application (idempotent)
      responses:
        "200":
          description: Favorite state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FavoriteResult" }
        "404":
          description: application not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/applications/{id}/avatar:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    get:
      operationId: getApplicationAvatar
      tags: [applications]
      summary: Stream avatar image bytes
      responses:
        "200":
          description: Image bytes
          content:
            image/png: {}
            image/jpeg: {}
            image/gif: {}
            image/webp: {}
        "404":
          description: avatar not set
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    post:
      operationId: uploadApplicationAvatar
      tags: [applications]
      summary: Replace avatar (multipart file; png/jpg/jpeg/gif/webp ≤ 2MB)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "400":
          description: field errors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FieldErrors" }
    delete:
      operationId: clearApplicationAvatar
      tags: [applications]
      summary: Clear avatar (falls back to emoji icon)
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
  /api/v2/applications/{id}/default-agent:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      operationId: setDefaultAgent
      tags: [applications]
      summary: Set as workspace main agent (chat + enabled binding only)
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "400":
          description: not eligible
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "403":
          description: only creator or staff
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    delete:
      operationId: unsetDefaultAgent
      tags: [applications]
      summary: Unset workspace main agent
      responses:
        "200":
          description: Updated (compact shape)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
  /api/v2/applications/{id}/attachments:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      operationId: uploadApplicationAttachment
      tags: [applications]
      summary: Pre-run attachment upload (stored in studio storage; the worker uploads to the provider with the user's UAT)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [image, file, feishu_doc, bitable]
                  default: file
                doc_url:
                  type: string
      responses:
        "201":
          description: Attachment created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuntimeAttachment" }
        "400":
          description: file or doc_url required / binding missing
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/DetailError" }
                  - { $ref: "#/components/schemas/BareError" }
                  - { $ref: "#/components/schemas/FieldErrors" }
        "502":
          description: attachment upload failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }

  # ───────────────────────────────────────────────────────── schedules ──
  /api/v2/schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List the caller's schedules (keyset pagination)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [all, running, paused, failed], default: all }
        - name: before_id
          in: query
          schema: { type: integer }
          description: Keyset cursor — return schedules with id < before_id.
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
      responses:
        "200":
          description: Schedule page (bare array; empty means no more)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Schedule" }
        "401":
          description: not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a schedule (owner = caller; enabled by default)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleUpsertRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "400":
          description: field errors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FieldErrors" }
        "401":
          description: not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/schedules/preview:
    post:
      operationId: previewScheduleRuns
      tags: [schedules]
      summary: Compute the next trigger times for a candidate trigger (no side effects)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleUpsertRequest" }
      responses:
        "200":
          description: Next trigger times (UTC ISO-8601), at most 3
          content:
            application/json:
              schema:
                type: object
                properties:
                  next_runs:
                    type: array
                    items: { type: string, format: date-time }
        "400":
          description: invalid trigger
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FieldErrors" }
  /api/v2/schedules/{id}:
    parameters:
      - $ref: "#/components/parameters/ScheduleId"
    get:
      operationId: getSchedule
      tags: [schedules]
      summary: Schedule detail
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "404":
          description: unknown or foreign schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a schedule (next_run_at recomputed from now)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleUpsertRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "400":
          description: field errors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FieldErrors" }
        "404":
          description: unknown or foreign schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule (occurrence/run/delivery history is kept)
      responses:
        "204":
          description: Deleted
        "404":
          description: unknown or foreign schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/schedules/{id}/enable:
    parameters:
      - $ref: "#/components/parameters/ScheduleId"
    post:
      operationId: enableSchedule
      tags: [schedules]
      summary: Enable a schedule (next_run_at recomputed from now; missed slots are not replayed)
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "404":
          description: unknown or foreign schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/schedules/{id}/disable:
    parameters:
      - $ref: "#/components/parameters/ScheduleId"
    post:
      operationId: disableSchedule
      tags: [schedules]
      summary: Disable a schedule (no new occurrences; already queued runs are kept)
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "404":
          description: unknown or foreign schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/schedules/{id}/run-now:
    parameters:
      - $ref: "#/components/parameters/ScheduleId"
    post:
      operationId: runScheduleNow
      tags: [schedules]
      summary: Trigger one real occurrence now (overlap policy applies; next_run_at untouched)
      responses:
        "202":
          description: Occurrence created and run queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduleOccurrence" }
        "404":
          description: unknown or foreign schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "409":
          description: overlap policy skip and a previous run is still active
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/schedules/{id}/occurrences:
    parameters:
      - $ref: "#/components/parameters/ScheduleId"
    get:
      operationId: listScheduleOccurrences
      tags: [schedules]
      summary: Execution history (keyset pagination, newest first)
      parameters:
        - name: before_id
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
      responses:
        "200":
          description: Occurrence page (bare array)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ScheduleOccurrence" }
        "404":
          description: unknown or foreign schedule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }

  # ─────────────────────────────────────────────────────────── runtimes ──
...
  /api/v2/runs:
    post:
      operationId: createRun
      tags: [runs]
      summary: Create a run (lazily creates the conversation when conversation_id is absent)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRunRequest" }
      responses:
        "201":
          description: Run created (status=queued)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "400":
          description: validation errors (bare array style)
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/BareError" }
                  - { $ref: "#/components/schemas/FieldErrors" }
  /api/v2/runs/{runId}:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: getRun
      tags: [runs]
      summary: Run detail (owner or staff only; unknown ids return 404)
      responses:
        "200":
          description: Run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "404":
          description: run not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/runs/{runId}/events:
    parameters:
      - $ref: "#/components/parameters/RunId"
      - name: after
        in: query
        schema: { type: integer, minimum: 0, default: 0 }
        description: Return events with sequence > after (ascending).
    get:
      operationId: listRunEvents
      tags: [runs]
      summary: Persisted run events (replay after reconnect)
      responses:
        "200":
          description: Events (bare array)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RunEvent" }
        "404":
          description: run not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/runs/{runId}/stream:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: streamRun
      tags: [runs]
      summary: SSE stream (DB replay from sequence 0, then Redis live frames)
      description: |
        `text/event-stream`. Frames:
        `event: run.event\ndata: {"run_id","sequence","event_type","payload"[,"created_at"]}\n\n`
        Keepalive comment `: keepalive` after 15s of silence.
        Closes after replay when the run is already terminal, or after the
        first terminal live event.
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          description: run not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
...
  /api/v2/runs/{runId}/artifacts:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: listRunArtifacts
      tags: [runs]
      summary: Artifacts of a run (created_at ascending)
      responses:
        "200":
          description: Artifacts (bare array)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RunArtifact" }
        "404":
          description: run not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
...
components:
...
  schemas:
...
    SimpleError:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Tokens:
      type: object
      properties:
        access: { type: string }
        refresh: { type: string }
...